requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.14.56
botocore==1.17.56
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
Django==3.1
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.11.1
docutils==0.15.2
flower==0.9.5
humanize==2.6.0
idna==2.10
jmespath==0.10.0
kombu==4.6.11
Pillow==7.2.0
prometheus-client==0.8.0
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
vine==1.3.0
